Moved Dulmage-Mendelsohn decomposition into new decompositions package
